backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "firebase-admin>=7.4.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]
